requirements.txt
Flask==3.1.2
pymongo==4.15.1
playwright==1.55.0
beautifulsoup4==4.14.2
python-dotenv==1.1.1
twilio==9.8.3
requests==2.32.5
lxml==6.0.2
